[CP-3275] Typography modifiers moved to the typography catalog
