+++ libs/generic-view/ui/src/lib/typography/apply-text-transform.ts
