+++ libs/generic-view/ui/src/lib/typography/format-bytes.ts
